requirements.txt
transformers>=4.28.0
torch>=2.2.0
Pillow>=9.0.0
gradio>=4.44.0
streamlit>=1.18.0
